requirements.txt
aiohttp==3.6.2
aiohttp-wsgi==0.8.2
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
arcgis==1.8.0.post1
asn1crypto==0.24.0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.4
blinker==1.4
Brlapi==0.6.7
CacheControl==0.12.6
cachetools==4.0.0
certifi==2020.4.5.1
certstream==1.10
cffi==1.13.2
chardet==3.0.4
Click==7.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.12
debtcollector==1.22.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.21ubuntu4
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
fake-useragent==0.1.11
fhirclient==3.2.0
firebase-admin==4.0.1
flake8==3.7.9
Flask==1.1.1
Flask-API==2.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
future==0.18.2
GDAL==2.4.2
Geohash==1.0
geoip2==2.9.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.28.1
html5lib==1.0.1
httpie==1.0.3
httplib2==0.17.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.3.0
incremental==17.5.0
ipwhois==1.1.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
jeepney==0.4.3
Jinja2==2.10.3
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.1
keyring==21.2.0
keyrings.alt==3.1.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lerc==0.1.0
libvirt-python==5.0.0
louis==3.10.0
lxml==4.4.2
macaroonbakery==1.2.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
maxminddb==1.5.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
motor==2.1.0
msgpack==1.0.0
multidict==4.7.3
names-dataset==1.9.0
nbconvert==5.6.1
nbformat==5.0.5
netifaces==0.10.4
notebook==6.0.3
ntlm-auth==1.4.0
numpy==1.18.2
oauthlib==3.1.0
olefile==0.46
oslo.i18n==3.25.0
OWSLib==0.18.0
packaging==19.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
pbr==5.4.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.1
pipenv==2018.11.26
plotly==4.1.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.6.6
psycopg2==2.7.7
ptyprocess==0.6.0
pyasn==1.6.0b1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pydocstyle==5.0.1
pyee==6.0.0
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.34.0
PyHamcrest==1.9.0
PyJWT==1.7.0
pylint==2.4.4
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
pyproj==1.9.6
PyQRCode==1.2.1
PyQt5==5.12.3
pyRFC3339==1.1
pyrsistent==0.16.0
pyshp==2.1.0
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.7.3
python-debian==0.1.36
python-jsonrpc-server==0.3.2
python-language-server==0.31.2
python-twitter==3.5
pytz==2019.3
pyxdg==0.26
PyYAML==5.1.2
pyzmq==19.0.0
redis==3.3.11
rejson==0.5.2
reportlab==3.5.23
requests==2.23.0
requests-file==1.4.3
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
retrying==1.3.3
rfc3986==1.3.2
rope==0.14.0
rsa==4.0
scour==0.37
SecretStorage==3.1.2
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.0
shapely-geojson==0.0.1
simplejson==3.16.0
sip==4.19.18
six==1.14.0
snowballstemmer==2.0.0
soupsieve==1.9.2
system-service==0.3
systemd-python==234
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
tldextract==2.2.2
tornado==6.0.4
tqdm==4.40.2
traitlets==4.3.3
typed-ast==1.4.0
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==15.1.0
virtualenv-clone==0.5.3
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.48.0
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xkit==0.0.0
xmltodict==0.12.0
yapf==0.29.0
yarl==1.4.2
zipp==0.6.0
